DEBUG: See seismic_data "input" statements
Only difference between fwi.optimizer.param_groups and alan_optimizer.param_groups
is that the tensor being optimized in fwi is "parameter containing tensor==..." and
the tensor being optimized in alan_optimizer is just the tensor itself.

Make sure that this is what is needed or if we need to unwrap down to the tensor
level with the optimizer. I suspect that we are okay; a torch.nn.Parameter is a subclass
of tensor, and its "forward" function MATTERS in the optimization (since it affects the gradient computation).