BUG: Obscure bug on line 71 of misfit_toys/examples/ddp/driver.py
For some reason delay_init for Param breaks things and leads to
ValueError: bad value(s) in fds_to_keep error.

It works fine when we use ParamConstrained instead. This version of the
code is INTENTIONALLY BROKEN to demonstrate the bug. Just uncomment
the other vp_prmzt line and it will work fine.