FEATURE: Add example script for splines
Uses torchcubicspline; note that their documentation says that you input

(batchdim1, batchdim2, ..., timedim, input_dim)

However, this leads to issues if you want to use it as in our case with
Wasserstein-2 where our timedim is actually of shape (batchdim1, ... , batchdimN).
I tried to fix this by "flattening" along the input dimensions, and then just set the batch size to 1.
However, this seemed to evaluate ONLY ONE spline, not input_dim splines!
This is what led me to use unbatch_spline_eval...since this is what we need.
In our context, batch_dim = (n_shots, n_sources, n_receivers).
For each batch, we compute the quantile for the observed data and spline interpolate said quantile.
This would be fine to just use batch_dim = true_batch_dim, and input_dim=1, but the issue is that we evaluate each quantile along a DIFFERENT time series!
In other words, the renormalized CDF of the synthetic data.
There might be a way to achieve this without using unbatch_spline_eval, but I couldn't figure it out.
This fix is good enough for now.