FEATURE: Add simple.py for exponential distribution
This is (possibly) a better class to unit test W2 with, given that
we have an easy closed form solution for every term in the loss AND
the gradient for W2.

Note that some filtering needs to take place to deal with the tails of the distribution.
This is less of an issue with the seismic data.
However, notice how in the supplied gif that even though the quantile
functions seem to be a great match on the regular grid, they seem
significantly far off under composition with the CDF.

However, we seem to get the loss landscape we seek (~roughly proportional...see bottom left figure).
This is because OT is a smoothing operation and thus the fact that we miss local behavior "cancels out" after we integrate.
But keep this composition thing in mind, as it leads to significant error.
Seismic data seems to actually be a little less sensitive to this than the exponential distribution.

Some ideas of how to handle this:
(1) In theory, you can create yet another spline interpolation of the derivative of the quantile.
    This may lead to some more accurate results.
    Furthermore, if we want to integrate the Hessian, we would need to do this anyway.

(2) The issue lies at the boundary...threshold the input to quantiles 0.05 and 0.95, say, as this will restrict the input data into the range where the quantile function is smooth.

(3) Compose with a smoothing operation such as a sigmoid.