DOCS: Format looks roughly correct
TODO: I believe when we start CI again, we will have
a build issue due to a changed path. In particular,
the _leaderboard directory is now completely gone,
so the leaderboard update might incorrect for a few pushes
due to this path change and fix after that. Or there may
be a bug since this is currently untested.

TODO: Consider allowing for a collapsible toc index.

[skip ci]