DOCS: Remove empty case leaderboards
TODO: This simplifies the toctrees significantly (for now).
We need to populatte the site with more test cases.

TODO: The timestamps are not being preserved due to the
different file systems. Say January 1 gets pushed to the repo.
Then when we pull, a new file named something without a timestamp
is there. This is a problem because the timestamp
is currently generated by the file system. We need to
generate this timestamp information directly on the local files
and then simply compare with the pulled cloud version so that
we have an accurate timestamp that doesn't get destroyed by
continuous integration.