FIX: __call__ resolution
TODO: Allow the resolution mechanism to be overriden by the user.

The user simply specifies

resolution_order
  - self_key: slf_whatever1
    call_key: __call_whatever1__
  - self_key: ...

  and then the main script just looks like

for e in c.resolution_order:
   c = resolve(c, self_key=e.self_key, call_key=e.call_key)

This is both less verbose, and a bit clearer honestly.
This resolution order will often be exactly the same, but
then the user can see exactly what I'm trying to do here (arguably it's a bit more confusing also since more is abstracted away from python and into the yaml...consider this carefully).