BUG: Data deployment is massively flawed
For deepwave_example, you need to actually
run the forward solver AGAIN, right?

As of right now, all you do is downsample the model parameters
and observed data. However, this is flawed.
In the original Marmousi case, you have 2301x751 grid and 751 time steps.
Now, you get a forward solve which is fine and dandy.
But think about the physics!
Your observed data there has the upper left hand corner NOWHERE NEAR
the right boundary, meaning that error from that PML is essentially nonexistent.
So if you are now going down to a 600x250 grid with 300 timesteps, you cannot just downsample
the data there. You need to actually run the forward solver again.
This is why you get that small discrepancy in your FWI resullts.
I believe a bit of it is randomness in the optimization, but this is still flawed!