FIX: Marmousi2 data extraction
Previously, you were not extracted both components of the observed data.
Now you are.

BUG: Note the comment in the DataFactory class for the data creation callback.
For the excluding directories, the edge case of when an exclusion directory is a substring
of one that one wishes to include will cause the one you want to include to actually be excluded.
For now, I have just refactored marmousi2 -> marm2 rename to avoid this issue, but this should be rectified in the future.